👌 IMPROVE: refacto trait GraphQl client
